1209. Remove All Adjacent Duplicates in String II
