+++ src/test/kotlin/me/bossm0n5t3r/leetcode/removeAllAdjacentDuplicatesInString2/RemoveAllAdjacentDuplicatesInString2Test.kt
@@ -1,10 +1,10 @@
-package removeAllAdjacentDuplicatesInString2
+package me.bossm0n5t3r.leetcode.removeAllAdjacentDuplicatesInString2
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 internal class RemoveAllAdjacentDuplicatesInString2Test {
-    private val removeAllAdjacentDuplicatesInString2 = RemoveAllAdjacentDuplicatesInString2()
+    private val removeAllAdjacentDuplicatesInString2 = RemoveAllAdjacentDuplicatesInString2.Solution()
 
     private data class RemoveAllAdjacentDuplicatesInString2TestData(
         val s: String,
